crossmint: change to type
