+++ packages/common/base/src/apiClient/CrossmintApiClient.ts
@@ -1,7 +1,7 @@
-import { Crossmint } from "@/Crossmint/Crossmint";
 import { environmentToCrossmintBaseURL } from "@/apiKey/utils/environmentToCrossmintBaseURL";
 import { ValidateApiKeySuccessResult, validateAPIKey } from "@/apiKey/validateAPIKey";
 import { ValidateAPIKeyPrefixExpectations } from "@/apiKey/validateAPIKeyPrefix";
+import { Crossmint } from "@/types/Crossmint";
 
 import { ApiClient } from "./ApiClient";
